requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
aiofiles==23.2.1
    # via chainlit
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.10.8
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==3.7.1
    # via
    #   asyncer
    #   httpcore
    #   openai
    #   starlette
    #   watchfiles
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
    #   redis
asyncer==0.0.7
    # via chainlit
attrs==24.2.0
    # via aiohttp
bidict==0.23.1
    # via python-socketio
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
chainlit==1.1.404
    # via -r requirements.in
charset-normalizer==3.3.2
    # via requests
chevron==0.14.0
    # via literalai
click==8.1.7
    # via
    #   chainlit
    #   uvicorn
dataclasses-json==0.5.14
    # via
    #   chainlit
    #   langchain-community
deprecated==1.2.14
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-semantic-conventions
distro==1.9.0
    # via openai
exceptiongroup==1.2.2
    # via anyio
fastapi==0.110.3
    # via chainlit
filetype==1.2.0
    # via chainlit
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
googleapis-common-protos==1.65.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
greenlet==3.1.1
    # via sqlalchemy
grpcio==1.66.2
    # via
    #   grpcio-tools
    #   opentelemetry-exporter-otlp-proto-grpc
    #   qdrant-client
grpcio-tools==1.62.3
    # via qdrant-client
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
    #   wsproto
h2==4.1.0
    # via httpx
hpack==4.0.0
    # via h2
httpcore==0.17.3
    # via httpx
httpx[http2]==0.24.1
    # via
    #   chainlit
    #   langsmith
    #   literalai
    #   openai
    #   qdrant-client
hyperframe==6.0.1
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.4.0
    # via opentelemetry-api
jiter==0.5.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.0
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.3.0
    # via -r requirements.in
langchain-core==0.3.1
    # via
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   langchain-qdrant
    #   langchain-text-splitters
langchain-openai==0.2.0
    # via -r requirements.in
langchain-qdrant==0.1.4
    # via -r requirements.in
langchain-text-splitters==0.3.0
    # via langchain
langsmith==0.1.121
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
lazify==0.4.0
    # via chainlit
literalai==0.0.607
    # via chainlit
marshmallow==3.22.0
    # via dataclasses-json
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
nest-asyncio==1.6.0
    # via chainlit
numpy==1.26.4
    # via
    #   chainlit
    #   langchain
    #   langchain-community
    #   qdrant-client
openai==1.51.0
    # via langchain-openai
opentelemetry-api==1.27.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   uptrace
opentelemetry-exporter-otlp==1.27.0
    # via uptrace
opentelemetry-exporter-otlp-proto-common==1.27.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.27.0
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.27.0
    # via opentelemetry-exporter-otlp
opentelemetry-instrumentation==0.48b0
    # via uptrace
opentelemetry-proto==1.27.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.27.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   uptrace
opentelemetry-semantic-conventions==0.48b0
    # via opentelemetry-sdk
orjson==3.10.7
    # via langsmith
packaging==23.2
    # via
    #   chainlit
    #   langchain-core
    #   literalai
    #   marshmallow
portalocker==2.10.1
    # via qdrant-client
protobuf==4.25.5
    # via
    #   googleapis-common-protos
    #   grpcio-tools
    #   opentelemetry-proto
pydantic==2.9.2
    # via
    #   chainlit
    #   fastapi
    #   langchain
    #   langchain-core
    #   langchain-qdrant
    #   langsmith
    #   literalai
    #   openai
    #   pydantic-settings
    #   qdrant-client
pydantic-core==2.23.4
    # via pydantic
pydantic-settings==2.5.2
    # via langchain-community
pyjwt==2.9.0
    # via chainlit
python-dotenv==1.0.1
    # via
    #   chainlit
    #   pydantic-settings
python-engineio==4.9.1
    # via python-socketio
python-multipart==0.0.9
    # via chainlit
python-socketio==5.11.4
    # via chainlit
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
qdrant-client==1.11.2
    # via
    #   -r requirements.in
    #   langchain-qdrant
redis==5.1.1
    # via -r requirements.in
regex==2024.9.11
    # via tiktoken
requests==2.32.3
    # via
    #   langchain
    #   langchain-community
    #   langsmith
    #   opentelemetry-exporter-otlp-proto-http
    #   tiktoken
simple-websocket==1.0.0
    # via python-engineio
sniffio==1.3.1
    # via
    #   anyio
    #   httpcore
    #   httpx
    #   openai
sqlalchemy==2.0.35
    # via
    #   langchain
    #   langchain-community
starlette==0.37.2
    # via
    #   chainlit
    #   fastapi
syncer==2.0.3
    # via chainlit
tenacity==8.5.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tiktoken==0.7.0
    # via langchain-openai
tomli==2.0.2
    # via chainlit
tqdm==4.66.5
    # via openai
typing-extensions==4.12.2
    # via
    #   asyncer
    #   fastapi
    #   langchain-core
    #   multidict
    #   openai
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspect
    #   uvicorn
typing-inspect==0.9.0
    # via dataclasses-json
uptrace==1.27.0
    # via chainlit
urllib3==2.2.3
    # via
    #   qdrant-client
    #   requests
uvicorn==0.25.0
    # via chainlit
watchfiles==0.20.0
    # via chainlit
wrapt==1.16.0
    # via
    #   deprecated
    #   opentelemetry-instrumentation
wsproto==1.2.0
    # via simple-websocket
yarl==1.13.1
    # via aiohttp
zipp==3.20.2
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
